+++ src/hooks/setInterval.tsx
@@ -1,0 +1,21 @@
+import { useEffect, useRef } from "react";
+
+export function useInterval(callback: any, delay: number | null) {
+  const savedCallback: any = useRef();
+
+  // Remember the latest callback.
+  useEffect(() => {
+    savedCallback.current = callback;
+  }, [callback]);
+
+  // Set up the interval.
+  useEffect(() => {
+    function tick() {
+      savedCallback.current();
+    }
+    if (delay !== null) {
+      let id = setInterval(tick, delay);
+      return () => clearInterval(id);
+    }
+  }, [delay]);
+}
